Adding missing BLAS and LAPACK TPLs for PETSC (VRI Kanban 2954)
With the current TriBITS system, you must list all of the TPLs that need to be
linked in.  While TriBITS handles the ordering of the TPLs is does not
currently manage the metadata needed to know how one TPL depends on another.
Therefore, if you list an intermediate TPL and not its upstream TPLs, they can
be missing, or be in the wrong order, on the link line.  This is a weakness of
the TriBITS TPL system that must be fixed at some point.

The issue is that some MPACT subpackages were listing a dependence on PETSC
but not BLAS *and* LAPACK.  This resulted in a link failures on the
westinghouse machine for the test under:

  MPACT/MPACT_libs/Utils/unit_tests/testTPLPETSC

The LAPACK library was listed on the link line but is was listed *before* the
libpetsc.a library.  It would seem that library ordering still matters on some
systems.

Adding the missing BLAS and LAPACK libraries fixed the problem and is the
right thing to do.  It is not clear how MPACT was able to link with PETSC on
other systems.  It must just have been luck.