Adding %pathExists functionality to HDF5FileType.
This function will check whether the specified path exists in an HDF5 File.  Currently the only useful inquiry function I could create since checking if things are a group, dataset, or datatype is ugly at best.  So, atleast we can see if the path is currently in use or not, which is something.