Format edit of 1D arrays differently to save lines in output
The current way the PL dumps a 1D array is vertically.  This makes our
output file incredibly large and very frustrating to get to the output.
Now I've added a MAX_1D_LEN.  If the length is less than or equal to
that, it will print vertically, otherwise, it prints horizontally.
The current value of MAX_1D_LEN is 10.