Update to handle the closing of the HDF5 interface when more than 1 file exists.
Added a module counter and logical to determine when at least one HDF5 file type has been initialized.
Added a check in the clear routine to determine if the file being closed is the last HDF5 file that has been initialized.  If so, it will close the HDF5 interface, instead of it being called everytime an object is cleared, which resulted in HDF5 error messages.