Fixing bugs in parameter list has/get methods
The get method allows for partial matching, however if a
parameter was identified as "a->b" in the list, a parameter having
name "b" under any parameter named "a" would be returned. This meant
that we would incorrectly identify parameters.

Instead of finding "any b under any a in the list" we now find
"any a->b" under the list. So the searching/matching is done for
the exact address given. This was a major bug in the %get method that
is used widely by other methods.

In addition to this, the "has" bug where one might need to use
a list container before properly searching has now been fixed.
(Ticket #507 at UM).