Slight mod to defineQuadraticEdge.
We realized we needed a way to determine the "inflection"
of the curve, in the case of a semi-circle. The solution is
that the radius of for the curve may be negative ONLY in the
case of a semi-circle to determine orientation of the curve
(e.g. additive or subtractive to a volume). The PolygonType needs
to know this.