Finishing Geom_Poly %inside and %intersectLine routines
The geom_circ arc intersection routine did not take into account the
circle's centroid when checking the arc edges.  That issue is fixed now.
Test will be updated shortly.

The %inside and %intersectLine are under test.  %inside is tested
thoroughly, and there may be room for more testing of the %intersectLine
functionality.  There is enough testing for now to push out for
GenPinMeshType development.