Fixing the %set routine.
Found an inconsistency depending on if the graphtype's starting point is
in the lower left with the second point directly above, or if there is no
point above the starting point.  The %getCCWMostVert would return
different vertices depending on if the above case was encountered.

Logic was added to test if the vertex returned from either %getCCWMostVert
or %getCWMostVert is above the starting point and if so, to use that
point.  The vertices within the PolygonType are still numbered in
clock-wise fashion.  In updating the %inside test, this issue was
discovered.