Fixing intersectArcLine case.
The case where the arc crosses the polar coordinate line theta=0 would
incorrectly clear the valid intersection points found on that arc
interval.

Logic was added to make sure that thetastp > thetastt, and the calculated
theta for the intersection point would be incremented by two pi if it was
above the polar coordinate line theta=0 and if thetastp was modified as
well.

Unit test case was added.