Adding Andrew's fix for VTK_edit.  Fixes for pin mesh reference frame.
The pin geometry's reference frame for 0,0 is the lower left of the pin
cell, while the pin mesh is in the center of the pin.  The geometry
objects that are created now take this shift into account.

The geom_poly subtract subregion will now behave correctly when it needs
to remove overlapped regions.  The test for this case was updated.