Fixing the initialization tests for the control rod geometry.
Some of the code additions made from similar necessary checks done in
setting up the PWR control rods required some changes to the error
messages.

The names of some of the messages were also changed in light of moving
from an insert type to  a pin geom type for the rodlet specifiation.

The dz and startpos attributes from the control rods were added to the
control blades as well.  The exception cases for these cases are not yet
tested.