Changing the name of the XS Macro in %editXSLib.
A bug was exposed in nFields.  Added failing cases to the unit test,
commented them out.

This bug was causing the getField routine to not work properly for a
string with multiple fields and quotes, so as a work around, the material
name is printed as the comment, and the macro name is just a indexed name.