Updating XML material processing
Description:
Added a new section of code to organize materials by their block.

In doing this there are also checks for XS set names that are also material names.
The XS set names are appended to have the block name information if that is the case.

Then the material names themselves are appended with their block names.  This two
step process ensures consistency of names when materials are being mixed.

The interface to setupMats was changed, as well as adding two other intermediary
routines.

The unit test has been updated to reflect the changes.

Fixing a GNU bug in the overloaded geom graph assignment.
UM Ticket # - N/A
CASL Ticket # - PHI #3776