Added strarrayeqind function.
Description:
This function behaves similarly to the strarraymatchind function.
An array of strings is searched for the given pattern, and this function will
return the index of the pattern in the array only if the pattern is matched
exactly.  Strarraymatchind will return the index if the function strmatch
returns true.

This routine is needed for tighter ZZAAA and XSSet name checking.

UM Ticket # - N/A
CASL Ticket # - PHI #3776