Fixing Intel bugs and overloaded assignment operation bugs.
Description:
In Intel I was having issues with the PL's getting some of the correct values.  I changed
the %get calls to use the absolute paths for the different blocks that we scan which appears
to have fixed the issue.  This fix was also done for other sublists as well, like cells and
cellmaps.

Ran into another really freak issue as well with the overloaded assignment of char_to_*
routines.  We had a variable in an input file with a ridiculous amount of digits, which
ran into the upper limit of the formatting for the READ statement.  This caused the value
to be truncated in some instances, which produced the wrong solution as a result.

The fix for this was to make the formatting general.  So the format statement now uses
the length of the character string that is passed in as the limit for the formatting statement.

UM Ticket # - N/A
CASL Ticket # - N/A