Modifying ParameterList editToXML to preserve original formatting
Description:
When exporting the file and writing doubles or arrays of doubles, the
format was questionable. Now if the double values are read in from a file
and then written back out again, if the original values did not change,
the value in the file does not change. To do still is use significant
figures to format the output value when it does change.

CASL Ticket # - #4189