Fixing ParameterLists::getSubParams for Intel 13 compiler bug
Description:
The bug relates to the association of the arguments in the ASSOCIATED
intrinsic function and whether they are the POINTER or TARGET. In
one instance the function returns the correct result, and in the other
instance it does not. Later versions of Intel and gfortran compiler return
the same result regardless of the order/association of the dummy arguments.

CASL Ticket # - #4200