Bug fix to ParamType add when adding parameters
Description:
There was a bug when trying to add a parameter list as a sublist in
another parameter list. If the name of the list being added was longer
than the argument of the address to where the sublist should be added,
then the sublist name would be truncated. This was fixed by changing some
internal variables to StringTypes.

CASL Ticket # - N/A