Fixing CR Movement init not finding materials properly
Description:
The loop that fills the material ID array was suffering from floating
point comparison problems when a heterogeneous rod aligned with a mesh
boundary.  It stopped one axial plane too soon, and when it went to go
find the inserts that were in that plane, it did not find them, nor their
material IDs.  Loosened the tolerance on the comparison, and it now
initializes properly.

CASL Ticket # - PHI #3840