Fix to getRealFormat in IO_Strings for Intel
Description:
Intel apparently doesn't have problems with numbers defined like:
"3.2e". It assumes this is "3.2e0", this caused the routine to fail on
a degenerate case, and produce a bad format string, when it should
produce an empty format string. Was able to find another way to
catch this case, since the Intel Fortran I/O routines do not
think this is an issue.

CASL Ticket # - N/A