Nasty Intel bug fix in new Parameter list routine
Description:
Previously a list iterator was added to the ParamType. This routine
was failing in Intel whenever the pointer argument was NULL.

To fix it, I switched the pointer/target arguments and created a
local temporary pointer variable to pass instead of the dummy arg
that was defined as a target.

CASL Ticket # - #4189