Adding check for ls function to check if the path is a group.
Description:
Was running into issues where this function would be called on a dataset
and not a group.  Added a check to the routine so that it returns an
allocated array of the sub-objects if the path entered is a group.  It
will return a deallocated array if the path is not a group.

CASL Ticket # - PHI #4314