Modifying the %pathexists routine to work for the '/' path.
Description:
It would seem that the root path was not registered as existing when
using the h5lexists_f routine.  This affected the isgrp not thinking it
was a group, which in turn affected the ls function not returned the
list of objects.

The fix was to check if the path passed in is in fact length 1 and the
first character equals '/'.

CASL Ticket # - PHI #4314