Making read_st0_helper actually read scalars since we are writing them.
Description:
The logic for the read_st0_helper was duplicated from the
read_st1_helper routine.  In it, there is logic to determine whether the
dataset to be read is a character array or a string.  This necessitated
the creation of read_ca0.  It was patterened after read_ca1.

This effort was primarily done to support the reading of logicals into
the parameter list object.

CASL Ticket # - PHI #4314