Refactor function pointer into encapsulated class which is extended
Description:
Previous design was causing issues with Intel, having an abstract interface
with a function pointer passed as an argument.  There was also an unresolved
issue with getting data (constants) passed into the function which is resolved
by having an extended class.  Overall, the design is considerably better.

CASL Ticket # - 4579