Fix setting number of OpenMP threads in PE
The settings in the OMP parallel environment were getting squashed by the call
to omp_set_num_threads(1) in the initialization of the Trilinos solvers. This
call was there because in the absence of a defined OMP_NUM_THREADS environment
variable, the Trilinos solvers would use all cores on a system. Manually setting
OpenMP to use on thread in the Trilinos initialization affected the result of
omp_get_max_threads, which was being treated as the maximum allowable number of
threads, effectively limiting the number of threads to 1 for all cases.

This commit changes the thread setting logic to:
 - Make sure that the requested number of threads is <= to the number of
   physical cores, otherwise raise a warning.
 - Call omp_set_num_threads() to the maximum number of threads requested over
   all instances of the OMP_EnvType
 - If the requested number of threads is zero, use omp_get_max_threads() to as
   the number of threads. This is not the same as omp_get_num_procs(); read the
   specification for details. Essentially this sets the number of threads to the
   OpenMP default value, which is usually the value of OMP_NUM_THREADS, if
   defined, or the result of omp_get_num_procs().

Description:

CASL Ticket # - N/A