Changing some arrayutils interfaces, removing optional args.
Description:
Some optional arguments can be good, too many are a pain.  Reducing the
number of them, making some previous ones required, including the DELTA.
Removing the XI argument all together because it is possible to prepend
an array variable using the (/SRK, var/) notation.  This makes that
optional argument redundant.

Adding the TOL optional argmuent to more of the array utils routines
since it will likely be needed to resolve the control rod bug.

These changes are being made in preparation for the addition of the TOL
arguments in the control rod geom.

CASL Ticket # - PHI #3840