Adding tol to some routines, removing tol limit.
Description:
Some of the arrayutils routines did not use the tolerance input option.
That is fixed for findUpBound and findLowBound, and by association
findEleHtAbove and findEleHtBelow.

There was a range for valid tolerances in the findIndex routine.  The
upper bound check on this has been removed, so the tolerance can be as
coarse as needed.

Adding the tolerance value from parameter TOL_GEOM_cm to all of the
calls to findIndex, findEleHtAbove and findEleHtBelow.

CASL Ticket # - PHI #3840