Use the parameter list initializer for Anasazi
Description:
This adds a preprocessor branch to create and use a ForTeuchos parameter list to
configure the Anasazi solver in the EigenvalueSolverTypes. For now, it is just
manually setting some of the defailts that live in the Trilinos solver wrappers
already, but serves as a good place to customize later.

CASL Ticket # - 4169