Fixing hdf5 file exception handler behavior after close
Description:
When calling the close operation for an HDF5 file the file's
exception handler is set to not stop on errors. However,
the stop on error property is not reset after the file is closed.
This causes problems when an HDF5 file is open and closed and
reopened during execution.

CASL Ticket # - #4776