Closing HDF5 id's that were left open.
Description:
This commit is comprised of two general changes.
The first is that there were some general utility routines that did not
properly close out their HDF5 id's after they were opened, created, or
getted.

The second was a structural issue in all of the writes and reads.  In
both cases, each has it's own pre and post read and write.  The post was
only called if there were no errors in the pre call.  It was possible
however to open some HDF5 objects in the call to pre*, have an error in
an object afterwards, and therefore the first opened objects would not
have been closed.  The structure of all of the reads and writes was
changed to always call pre and post regardless of the error codes.

Ultimately, this is the bug encountered in 4914.

CASL Ticket # - PHI #4914