Changing HDF5 file closure behavior, resetting closed file id
Description:
Changing the HDF5 file closure degree to SEMI.  This will prevent future
development from leaving HDF5 identifiers opened, as it will fail when
it tries to close the HDF5 file.

When the file is closed, the HDF5 file id is reset to 0 since a new
value is given each time it is opened.  This change is more for
bookkeeping, as the intents of the open and close routine would simply
overwrite the old file id.

CASL Ticket # - PHI #4914