Stop deallocating LU factorization all the time
Current implementation would clear and deallocate the preconditioner whenever the matrix
is changed or the solver failed.  This ends up spending significant time in parameter list
and allocation time for mongoose which is solving a temperature equation many many times
per solve.  The changes represent about an order of magnitude speedup in mongoose.