Added back missing variables, restoring test features for gcc
Variables were removed because GCC said they were unused.  Upon further inspection, there
was an ifdef on if the compiler was GFORTRAN that disabled pieces of the test.  It appears
that the feature that was being tested for intel only has been added into newer version of
gcc including 4.8.  The variables were added and the tests now run under both intel and gcc.