Update sundials ode solver to support multiple ODE solver objects
The sundials fortran interface effectively defines a singleton object
for the ode solver.  Since multiple ODE solvers are needed in Mongoose
in order to solve multiple pins at once.  This requires a lot of sundials
data to be global across all ODE solver types so module data structures are
used to hold a few logicals and data arrays that sundials requires.  Note that
this fix will only work if all ODE solvers have the same number of unknowns.
A better and more permanent fix would be to use the PETSc wrapper for
sundials or directly wrap the C interface for sundials if multiple size ODE solvers
are needed.