Added Recursive Inertial Bisection (RIB) method to partitionGraph, and some code reduction routines.
Description:
Added the Recursive Inertial Bisection (RIB) method to the PartitionGraphType. This method requires
coordinates on the graph, and uses inertial moments to bisect the graph.

Because the 3 recursive methods RSB, REB, RIB have similar structure, 2 additional routines were
added to reduce code duplication. These routines are:
1. detGroupSize - Determines the expected size of each group
2. recursivePartitioning - This routine takes in lists of each group (L1,L2) and refines them,
    redistributes groups if necessary, and calls the partition routine recursively.

A 3rd helper routine was added to determine order from a set of eigenvectors. In the RSB method,
the actual eigenvectors (Fiedler vector, or larger eigenvectors) are used to determine order. In
the RIB method the dot-product of coordinates with the eigenvectors is used. This routine, allows
one to recursively use a list of eigenvectors to determine order. It uses the first eigenvector,
then in the event of ties, the next eigenvector.

Cleaned up and reduced a lot of code in the unit-test as well. Almost all the partition checks were
the same 50 lines of code. Threw this into a routine and it is now called to check each partition
sub-test.

CASL Ticket # - #5029