This avoids a hang in MatTranspose
This issue is specific to MPICH implementations of MPI and
specifically a bug in IBarrier. The alternative is to have
PETSc formulate two-sided communication differently. In this case
the other option is to use an all reduce. Naturally this likely will
not scale as well.

Longer term we should consider updating to a newer version of MPICH
where this has been fixed.