This is probably a better fix for the DA32 filetype
Description:
It would appear that the TRANSFER intrinsic does not pad with 0 bits.
When converting a character into an integer, if the character is say
length 5, and characters are are a single byte and integers are 4 bytes
Then you need 2 integers to store the 5 characters, however, one
integer only gets one byte and therefore has 3 uninitialized bytes.

This may have been the root of the problem in this test. TRANSFER
won't pad 0 valued bytes so integers with 4 bytes had only 1 initialized.
Solution was to pad the string with spaces for the call to TRANSFER.

Fixed a valgrind defect and test passes in docker.