Create subroutine fillInterpMats in LS_MG
This subroutine is designed to fill out the interpolation matrices from the
information stored in a MultigridMeshStructureType object (see previous commit)

Added a new solve subroutine for LS_MG that does a quick check to see if MG
has been setup before calling its parent solve subroutine.

I am using 1-indexing for the multigrid meshes, but PETSc uses
0-indexing.

xyzMap is good for keeping track of where the cells are and it is very useful
for determining which cells to keep when coarsening.

For all grid points of interpolation degree higher than 0, childIndices
stores indices of the current grid rather than the coarser grid.  One can
recursively trace through the childIndices of its neighbors to find the
corresponding coarse grid points.  Grid points with interpolation degree 0
have a size-1 childIndices array that stores the coarse grid index of the
point it injects into.

The multigrid solver can support multi-equation (e.g., multigroup)
problems.  A helper function was added to extract the interpolation weights for
a fine mesh point by traversing the childIndices property of MultigridMesh
elements.