Create SmootherTypes module
Unit tests were also created.

We have implemented a general colored block Jacobi solver, created a new index
list type to store lists of indices for each color, and rearranged the KSP in
the unit test so that there's only one instance of it.

To use PCSHELL in petsc, we have to define several subroutines -- the most
important of which is PCApply, which takes in a vector x and returns a
preconditioned vector y.  This preconditioning (smoothing) step is where the
CBJ smoother will act on the preconditioner vector y.  However, it needs
access to the coloring schemes (and possibly the LU factorizations of the
blocks) which are/will be saved in the SmootherTypes.  Thus, a module-scope
list is needed, and this list is managed by various manager subroutines that
can be called by external modules.

A smootherID is assigned to PC context in PETSc.  Basically, each PETSc PC
instance will have an integer to store its smoother ID.  This took a lot of
jumping through hoops to do... The ID is also now in the name of the PETSc
preconditioner to distinguish the different preconditioners.