Implement colored, block solver in PCApply.
In order to complete the colored solver, num_colors KSP iterations must be run.
Each application of PCApply performs a sweep on one color.

A new attribute blockSolvers has been added.  This is an array of
LinearSolverTypes used to solve the blocks of the overall system.  For now,
the solver is defaulted to Futility's native LU.

Rework LSTypes_MG::setupPETScMG to use the setSmoother subroutine