Adding find/replace for C_NULL_CHAR in HDF5 read string routines.
Description:
The ultimate issue for the extra characters showing up in the state
variables was that they happened during a restart read.  The issue comes
from HDF5 reading in an empty string, where the C_NULL_CHAR is returned
if the string is empty.  Since we do not need that functionality, adding
a string find/replace after the data is read in so that C_NULL_CHAR is
replaced with ''.  Added scalar and array unit tests to check that this
change works properly.

CASL Ticket # - PHI #5241