Fixes to MeuLu parameter list
When we switched from ML to MeuLu, the ML Default Type parameter was
added to the incorrect parameter list, which is why it didn't cause issues.
This change gets rid of that parameter completely because multigrid algorithm
does the same thing.  sa: damping factor was also defined twice so the second
was not being used, yet the first value was incorrect and the second was correct
so I removed the first entry.

Ticket #3663