Change solver%interpMats_PETSc to pointer
This is in preparation to make it possible for multiple LS_MG objects to share
one set of interpolation matrices.

Also, we added a function to associate LS_MG%interpMats_PETSc pointer
and an accompanying unit test to make sure we can solve multiple
LS_MG instances with the same interpolation matrices.