Modification to MemProf to use better edit values
Description:
Using /proc/self/status VmRSS for the resident set size (RSS) versus
the VmSize parameter. The VmSize parameter will include memory from
shared libraries not explicitly allocated to physical memory or by
the program. This can lead to artificially high values that obfuscate
the actual memory usage.

The maximum memory (high water mark for the resident set size) is
also now utilized in the MemProf.f90 module. Previously it was just
the maximum value of VmSize.

Units are also now correct. Raw measurements in getSysProcInfo
are returned as "kilobytes" for peakUsage and "kB" for the RSS.

Generally:
kilobytes = 1000 bytes (or 1024 bytes, it is ambiguous)
kB        = 1000 bytes (may be used erroneously)
KB        = 1024 bytes (may be used erroneously)
KiB       = 1024 bytes (unambiguous, but not widely used)

The peakUsage is assumed to be 1024 bytes, and the VmRSS is assumed
to be 1000 bytes.