Add new getAll procedures for parallel vectors
Description:
The original getAll procedure will not work for a parallel vector because it
will cause an out of bounds error in PETSc.  To get all values at once, the user
needs to pass in the global indices of the values to get and they can only
pass  in indices that belong to the proc.  I added a new procedure, 'getSelected',
which can be used to do this, clarified that getAll should not be used for parallel
vectors, added tests for getSelected, and then added a new set of parallel vector
tests.