Fixing the case where a size 0 array is passed as an input argument.
Description:
I encountered a need for this to return zero if I pass in an array
subset making it a size 0 array (e.g. tmparray(1:0)).  Before this fix
it would return a 1, which can't be the case if there are no entries in
the array to be unique.  Added test cases for all of the unit test
types.

Also fixed a similar bug for getUnique.  Added tests for that as well.

CASL Ticket # - N/A