Adding verifyList and modifying the validateReq interface.
Description:
Adding an enumeration of the type of validation the validateReq will be
doing, since it is used for validating and verifying.

Changing the behavior of verify to be correct.  Before if the parameter
was not on both lists, it would not be verified and the call would still
pass.  Now, if the parameter exists on the required PL and isn't found
on the PL being verified, the match will return as a failure.  This
wasn't the case before and testXMLProc and testInputProc needed to be
updated accordingly.

CASL Ticket # - N/A