Adding an exception handler to the %verifyList interface.
Description:
It is a required input for the %verifyList procedure.  It is used
optionally in the validateReq_ParamType subroutine, since it is called
by several other procedures.

The unit test was updated for the interface change.  Both a temporary
exception handler and eParams were tested as input options.

CASL Ticket # - N/A