Updating PartitionGraph and metrics routine to account for weighting factor.
Description:
When calling the calcDecompMetrics routine, the %wts associated with
each vertex were already assumed to be normalized. This prevented the
ratio of process source regions to total source regions from being
calculated since they were normalized to 1.

A weighting factor was added to the type, the initialization parameter
list, and is used in the updated calcDecompMetrics routine to now return
the maximum ratio of process source regions to total regions.