Updating the %getString function.
Description:
Originally, the function would only return a string for scalar values,
and the developer was responsible for selecting the correct %getString*
subroutine depending on the parameter type being converted, whether a
scalar, 1-D, 2-D, or 3-D array.

This change allows the developer to use the scalar %getString function
for any dimension of parameter.