Multi-dimensional extrapolation added to interpolators module.
A 1-D extrapolation function is needed for the critical buckling calculation in MPACT.

Adds a 1-D extrapolation to the 1-D interpolation to the Interpolators module. If an optional argument is set to true, then it changes the behavior of the Interpolate routine to perform extrapolation instead of using the closest point in the table.

Extended 1-D extrapolation to bi-linear and tri-linear extrapolation. Corresponding unit tests added.