Changed needed for Windows support
Squash branch 'fixForWindowsHDF5' into 'master'

* Changed needed for Windows support

Writing string data to an HDF5 file on a Windows machine leads to
a segfault in the write_str procedures if the character array
size is defined in the procedure header.

This change is needed to allow string vectors to be written to the HDF5 file on Windows machines

For some unknown reason, applications using these procedures will crash when run on Windows.  Switching the character array to allocatable and explicitly allocating it to the correct size in the procedure fixes the issue.

See merge request https://code.ornl.gov/futility/Futility/-/merge_requests/375