Fix for newer versions of gcc
Squash branch 'fixesForGCC' into 'master'

* Fix for CI

* Fixes for when you have PETSc but not MPI

* Fully disable MemProf if disabled by user during configure

* Fix for newer versions of gcc

Newer versions of gcc compilers do not allow for implicit definition
of a function, so the ordering needed to change here.

These are some minor fixes for newer GCC versions

Implicit definition of a function in c++ is not allowed in newer versions of GCC, which just requires a simple reordering of a function here.  Additionally, modifications were made to `ifdef` out the other functions in the `MemProf` class so that none of the actions are performed when it is disabled.

See merge request https://code.ornl.gov/futility/Futility/-/merge_requests/376